[#171549167] Move the getUserGroups() method to the apim shared module
