+++ utils/apim.ts
@@ -1,4 +1,5 @@
 import { ApiManagementClient } from "@azure/arm-apimanagement";
+import { GroupContract } from "@azure/arm-apimanagement/esm/models";
 import * as msRestNodeAuth from "@azure/ms-rest-nodeauth";
 import { toError } from "fp-ts/lib/Either";
 import { TaskEither, tryCatch } from "fp-ts/lib/TaskEither";
@@ -28,4 +29,30 @@ export function getApiClient(
       ),
     toError
   ).map(credentials => new ApiManagementClient(credentials, subscriptionId));
+}
+
+export function getUserGroups(
+  apimClient: ApiManagementClient,
+  apimResourceGroup: string,
+  apim: string,
+  userName: string
+): TaskEither<Error, ReadonlyArray<GroupContract>> {
+  return tryCatch(async () => {
+    // tslint:disable-next-line:readonly-array no-let
+    const groupList: GroupContract[] = [];
+    const groupListResponse = await apimClient.userGroup.list(
+      apimResourceGroup,
+      apim,
+      userName
+    );
+    groupList.push(...groupListResponse);
+    // tslint:disable-next-line:no-let
+    let nextLink = groupListResponse.nextLink;
+    while (nextLink) {
+      const nextGroupList = await apimClient.userGroup.listNext(nextLink);
+      groupList.push(...nextGroupList);
+      nextLink = nextGroupList.nextLink;
+    }
+    return groupList;
+  }, toError);
 }
